\name{TuftePlot-package}
\alias{TuftePlot-package}
\alias{TuftePlot}
\docType{package}
\title{
This package aims to provide methods to create various graph strictly according to Tufte's principles.
}
\description{

}
\details{
\tabular{ll}{
Package: \tab TuftePlot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-24\cr
License: \tab \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dung Khuc

Maintainer: Dung Khuc <dung.khucquang@aalto.fi>
}
\references{
The Visual Display of Quantitative Information - Edward R. Tufte
}

\keyword{ package }
\seealso{

}
\examples{

}
